Bugfix: reduce CPU usage; better on_data handling
The while True loop in main() was calling queue.join() very rapidly,
causing full CPU usage when there was no reason. Added a sleep event to
the loop to prevent this.

The listener function on_data() ignored the fact that the token 'limit'
might appear somewhere in the tweet text or in some other value. To
prevent this, the JSON object is now always parsed from the data string
and then used to determine if the JSON object contains the relevant key
or not.