Switched from async to blocking call for stream
I can't think of a reason to have the main method open the stream in a
different thread rather than in the one that is running the main method.
The stream contents are processed by a different thread (or process in
the future). Also, and perhaps most importantly, if the stream has an
exception (somewhere in the tweepy library code), it is easier to handle
it with a try-except block rather than having an async thread die on its
own.