Fix: better handling of TCP/IP exceptions
Decided to try to handle top base-level TCP/IP exception classes as to
cast a wider umbrella. If we did not do this, the other option would be
to exhaustively list every possible type of TCP/IP exception that could
occur, which would be difficult to do until they happen at least once
(causing us to lose time reading the stream anyway).