Fix a UnicodeDecodeError causing data to be lost
Sometimes there is `UnicodeDecodeError` when trying to write a line to
the corpus file, but because it is unhandled it causes the writer-thread
to exit. In this case, the stream-queue continues to get filled, but the
additional incoming tweets can never be retrieved.

Therefore, we surround the line that writes the data to the file in a
try-except block which handles the Unicode error. This closes gh-1.